Updates in the positweb
Instead of using barcodes as Id that can be type-juggled like insanity by php, we are using incremental ids for find and a separate barcode_id find

Also, the return value for POSIT application wasn't the server id, so the meaning of synchronization got mangled. Somewhat fixed. I'm still not satisfied, but this is more correct.