This commit contains several revisions, mostly to PositMain and its menu/button interface.  In PositMain, the Adhoc Mode menu is updated to reflect whether RWG is running or not.  The Start/Stop menu is either enabled/disabled based on RWG status.  This does away with the need for the SharedPreference IS_ADHOC variable, which is no longer checked in FindActivity. Instead, RWGService.isRunning() is checked to see whether RWG is running and whether Finds should be sent via RWG.
In AdhocClient, I commented out the enableAdHocClient() method.  This
speeds up the starting of RWG considerably. This needs to be tested
with two rooted phones to verify that RWG still works. If RWG still
works, the AdhocClient code can probably be streamlined further.

Changes to other files were mainly cosmetic.