The server side changes that correspond to the refactoring of the POSIT synchronization activity.  Changes occurred mainly in the dao.php and (lesser so) in the api.php files.
THese files are very nicely organized and written.

One general feature that was added was to produce more Logging output.
Basically, each request from the client is logged in api.php and then
each function (worked on for this revision) makes a log entry at both
the beginning and end of the function.  This makes it a lot easier to
debug and develop.

One thing that could be improved is the error reporting and handling
between client and server.  We need a way for the client to pick up
from the server when a sync does not complete properly so that the
client can report this and so that the sync can be re-done or completed
in a subsequent attempt.