Major refactoring and revision of the following files:
SyncActivity
SyncThread
Find
FindActivity
ListActivity
PositDbHelper  ==> replaces MyDBHelper
Communicator

The changes seem to have removed (or drastically reduced) memory leaks
that were caused by various Db operations.  The policy attempted was to
stop passing Cursor back to calling methods such as when displaying
Finds in a list or Images in an adapter.  This was not entirely doable.
For example, SimpleCursorAdapter just has to have a Cursor--but it
seems to work well and not cause memory leaks.  All other cases were
modified.

There were also changes to SyncActivity, mostly having to do with
improvements to the UI, which does a better job of reporting what's
happening.  This Activity should be still further redesigned to put the
SYncing into a background thread.

The main changes were to improvements to ListFinds and FindActivity,
which now do the main functions of POSIT correctly, including updating
revision values, deleted values (Finds are now just marked 'deleted',
rather than removed from the Db), time stamped.  Many of these changes
required revisions to the Db both on the phone and on the server.

Comments are provided in the code.