Revised the TrackerActivity to run in a background thread.  The Tracker Activity is started from the PositMain menu.  When it is started it presents three menu items:  Start Tracking, Stop Tracking, and Return to PositMain.
Once the tracking thread is started the user can navigate away from the
Tracker using the 'Return to PositMain'.  From there, all of POSIT's
functions are available -- ListFinds, AddFinds, etc.  The back key can
be used to navigate back to the Tracker.

The Tracker can be stopped (putting it in the PAUSED) state and restarted
on the same 'expedition'.  The only way to end an expedition is to
use the 'back' key to exit the Activity.

Possible enhancements:
1)  Add a menu item to stop the thread (instead of using the current
stop item to pause it.
2)  Add a Network Connectivity Listener to the Activity so that it
does not attempt to send points to the serve if there is no network
connection.
3) Modify the functionality so that points are saved on the phone when
there is no network connectivity.
4) Add a feature that lets the user name the Expedition.
5) Add a feature that lets the user download and map expeditions
from this and other phones.